pyproject.toml
[tool.poetry]
name = "toggl-exporter"
version = "0.1.0"
description = "Project for exporting time entries once a day to Google Spreadsheet"
authors = ["Vladyslav Zlochevskyi <zlochvla@fel.cvut.cz>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
google-auth = "^2.32.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.136.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
toggl-exporter = "toggl_exporter.main:main"
